# Fix performance regression when indexing by Numpy arrays